(PC-5739) Move test to Home.spec.jsx
